Bypass existence and check for stable input for all StackedFuser classes
