+++ vllm/model_executor/models/transformers/fusers/qkv.py
@@ -19,9 +19,7 @@ from vllm.model_executor.models.transformers.fx_utils import (
     compile_forward,
     is_linear,
     recover_forward,
-    replace_expr,
     returned_linear,
-    self_call_and_refs,
 )
 from vllm.model_executor.models.transformers.utils import (
     log_replacement,
@@ -33,124 +31,6 @@ if TYPE_CHECKING:
     from vllm.config import VllmConfig
 
 logger = init_logger(__name__)
-
-
-def _is_none(node: ast.expr) -> bool:
-    return isinstance(node, ast.Constant) and node.value is None
-
-
-def _in_boolean_context(funcdef: ast.FunctionDef, ref: ast.expr) -> bool:
-    """Is `ref` used only for its truth value (a test, or a `not` operand)?
-
-    In these positions the object's identity never escapes, so a reference that
-    is always truthy can be replaced by `True`. `and`/`or` are excluded: they
-    yield an operand, so the module could escape (`x and self.<name>`)."""
-    for node in ast.walk(funcdef):
-        if (
-            isinstance(node, (ast.If, ast.IfExp, ast.While, ast.Assert))
-            and node.test is ref
-        ):
-            return True
-        if (
-            isinstance(node, ast.UnaryOp)
-            and isinstance(node.op, ast.Not)
-            and node.operand is ref
-        ):
-            return True
-    return False
-
-
-def _bypass_existence_guard(
-    funcdef: ast.FunctionDef, ref: ast.Attribute, name: str
-) -> None:
-    """Fold a guard on `self.<name>`'s existence to its constant value.
-
-    The fuser deletes `self.<name>` and binds only instances where it exists as a
-    truthy `nn.Linear` (guaranteed by the match, the cache key, and `validate`),
-    so a guard testing its presence is a fusion invariant. Two forms are folded:
-
-    - an identity `None` check `self.<name> is None` -> `False`,
-      `self.<name> is not None` -> `True` (either operand order). `==`/`!=` are
-      *not* folded: `is_linear` accepts `nn.Linear` subclasses, which may override
-      `__eq__`/`__ne__`, so equality is not guaranteed to track identity.
-    - a bare truthiness test (`if self.<name>:`, `... if self.<name> else ...`,
-      `not self.<name>`): the reference itself to `True`.
-
-    Any other surviving reference escapes the projection's value, which no longer
-    exists after fusion, so refuse rather than change semantics."""
-    for node in ast.walk(funcdef):
-        # `self.<name> is (not) None`, either operand order.
-        if not (isinstance(node, ast.Compare) and len(node.ops) == 1):
-            continue
-        (op,), (right,) = node.ops, node.comparators
-        if not isinstance(op, (ast.Is, ast.IsNot)):
-            continue
-        if (ref is node.left and _is_none(right)) or (
-            ref is right and _is_none(node.left)
-        ):
-            value = isinstance(op, ast.IsNot)
-            replace_expr(funcdef, node, ast.copy_location(ast.Constant(value), node))
-            return
-    # A bare truthiness test (statement `if`, ternary, or `not`); an `nn.Linear`
-    # is always truthy, so the reference folds to `True`.
-    if _in_boolean_context(funcdef, ref):
-        replace_expr(funcdef, ref, ast.copy_location(ast.Constant(True), ref))
-        return
-    raise ValueError(f"{name} is referenced outside an existence guard")
-
-
-def _base_name(node: ast.expr) -> str | None:
-    """The root `Name` of an attribute/subscript chain (`a.b[c]` -> `a`)."""
-    while isinstance(node, (ast.Attribute, ast.Subscript)):
-        node = node.value
-    return node.id if isinstance(node, ast.Name) else None
-
-
-def _rebound_names(region: list[ast.stmt]) -> set[str]:
-    """Names rebound outright within `region` (`x = ...`, `del x`).
-
-    These are `Name` nodes in a `Store`/`Del` context."""
-    return {
-        node.id
-        for stmt in region
-        for node in ast.walk(stmt)
-        if isinstance(node, ast.Name) and isinstance(node.ctx, (ast.Store, ast.Del))
-    }
-
-
-def _inplace_target(node: ast.AST) -> str | None:
-    """Base name `node` mutates in place, if any.
-
-    A write through the name (`x[i] = ...`, `x.attr = ...`) or an in-place method
-    call (`x.mul_(...)`) leaves the base name in a `Load` context, so it is not a
-    plain `Name` store (see `_rebound_names`)."""
-    if isinstance(node, (ast.Attribute, ast.Subscript)) and isinstance(
-        node.ctx, (ast.Store, ast.Del)
-    ):
-        return _base_name(node)
-    if (
-        isinstance(node, ast.Call)
-        and isinstance(node.func, ast.Attribute)
-        and node.func.attr.endswith("_")
-        and not node.func.attr.endswith("__")
-    ):
-        return _base_name(node.func.value)
-    return None
-
-
-def _mutated_names(region: list[ast.stmt]) -> set[str]:
-    """Names mutated in place within `region` (see `_inplace_target`)."""
-    return {
-        base
-        for stmt in region
-        for node in ast.walk(stmt)
-        if (base := _inplace_target(node)) is not None
-    }
-
-
-def _written_names(region: list[ast.stmt]) -> set[str]:
-    """Names rebound or mutated in place within `region`."""
-    return _rebound_names(region) | _mutated_names(region)
 
 
 @dataclass
@@ -223,18 +103,13 @@ class QKVFuser(StackedFuser):
         A projection may be guarded by an existence check -- a
         `self.<proj> is (not) None` comparison (e.g. Gemma 4's `v_proj`) or a bare
         truthiness test -- which is folded to its constant value (see
-        `_bypass_existence_guard`). The calls may sit in different branches, so the
+        `bypass_existence_guard`). The calls may sit in different branches, so the
         fused GEMM is inserted before the earliest of them, in the innermost block
         that dominates all three.
         """
         funcdef, fn = recover_forward(type(module))
         proj_names = (self.q_name, self.k_name, self.v_name)
-        calls = []
-        for name in proj_names:
-            call, refs = self_call_and_refs(funcdef, name)
-            for ref in refs:
-                _bypass_existence_guard(funcdef, ref, name)
-            calls.append(call)
+        calls = self._unguarded_calls(funcdef, proj_names)
         arg_dumps = {ast.dump(call.args[0]) for call in calls}
         if len(arg_dumps) != 1:
             raise ValueError("projection inputs are written differently")
@@ -268,18 +143,7 @@ class QKVFuser(StackedFuser):
         block = chains[0][depth - 1][0]
         indices = [chain[depth - 1][1] for chain in chains]
         insert_index = min(indices)
-        # The shared input must not be rebound or mutated between the insertion
-        # point and the last call, else the hoisted GEMM would read a different
-        # value. This covers `x = ...`, `x[i] = ...`, `x.attr = ...`, and
-        # in-place methods (`x.mul_(...)`); see `_written_names`.
-        arg_names = {
-            node.id
-            for node in ast.walk(calls[0].args[0])
-            if isinstance(node, ast.Name) and isinstance(node.ctx, ast.Load)
-        }
-        region = block[insert_index : max(indices) + 1]
-        if arg_names & _written_names(region):
-            raise ValueError("projection input is rebound or mutated before all calls")
+        self._check_input_stable(calls, block, indices)
 
         # q(x), k(x), v(x) -> q, k, v = qkv(x).split(qkv.output_sizes / qkv.tp_size, -1)
         self._splice_merged_split(funcdef, calls, block, insert_index)
